create files locally, add checks and convert final files to bw and bb

+++ coverage/issues.sh
@@ -13,10 +13,28 @@ if [[ "$#" -lt 5 ]]; then
 fi
 in_paf=$1
 name=$2
-in=${in_paf/.paf/}
 ver=$3
 platform=$4
 pattern=$5
+
+ln $in_paf
+in_paf=`basename $in_paf`
+in=${in_paf/.paf/}
+
+if [[ ! -s $in_paf ]]; then
+  echo "No $in_paf found. Exit."
+  exit -1
+fi
+
+if [[ ! -d $pattern/$ver || ! -s $pattern/$ver.exclude.bed || ! -s $pattern/$ver.telo.bed || ! -s $pattern/$ver.bed ]]; then
+  echo "No $pattern/$ver found. Exit. Run T2T-Polish/coverage/init.sh under pattern folder first."
+  exit -1
+fi
+
+if [[ ! -s $pattern/$ver.error.bed ]]; then
+  echo "No $pattern/$ver.error.bed found. Exit. Link the merged _only.bed file from Merqury as $pattern/$ver.error.bed."
+  exit -1
+fi
 
 ln -sf $pattern pattern
 
@@ -31,4 +49,12 @@ $tools/T2T-Polish/coverage/coverage_stat.sh $in.cov.wig pattern/$ver.exclude.bed
 
 # Collect low supportive regions
 $tools/T2T-Polish/coverage/low_support.sh $in_paf $ver $platform $pattern
+
+awk -v OFS='\t' -v FS='\t' 'NR==FNR { mapping[$1] = $3; next } \
+  {if ($3 > mapping[$1]) { $3=mapping[$1]; $8=mapping[$1];} print}' \
+  $pattern/$ver.bed $in.issues.bed > $in.issues.fm.bed
+
+module load ucsc
+wigToBigWig -clip $in.cov.wig pattern/$ver.fa.fai $in.cov.bw
+bedToBigBed -type=bed9 $in.issues.fm.bed pattern/$ver.fa.fai $in.issues.bb
